[FE] FEAT: AdminHook으로 코드 분할, 리팩토링 #985
